+++ ArtInk/ArtInk.Site/ViewModels/Response/InventarioProductoResponseDto.cs
@@ -8,11 +8,11 @@ public class InventarioProductoResponseDto
 
     public short IdProducto { get; set; }
 
-    public byte Disponible { get; set; }
+    public decimal Disponible { get; set; }
 
-    public byte Minima { get; set; }
+    public decimal Minima { get; set; }
 
-    public byte Maxima { get; set; }
+    public decimal Maxima { get; set; }
 
     public DateTime FechaCreacion { get; set; }
